Moved scale to standardization and added condition module
