+++ pymcr/__init__.py
@@ -3,4 +3,5 @@ from . import mcr
 from . import constraints
 from . import metrics
 from . import regressors
+from . import condition
 from . import rank
